[migrate] replace LeanCloud with Docker for deploying [add] VS Code configuration [optimize] Module path structure
