--- .github/workflows/build.yml
@@ -1,29 +1,0 @@
-name: Build TypeScript
-on:
-  push:
-    branches:
-      - master
-jobs:
-  build-and-deploy:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v2
-      - name: Use Node.js ${{ matrix.node-version }}
-        uses: actions/setup-node@v1
-        with:
-          node-version: ${{ matrix.node-version }}
-      - name: Build
-        run: |
-          npm install
-          npm run build
-          echo source/ >> .gitignore
-          echo \!dist/ >> .gitignore
-      - name: Release
-        uses: peaceiris/actions-gh-pages@v2.10.1
-        env:
-          PUBLISH_DIR: ./
-          PUBLISH_BRANCH: release
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        with:
-          forceOrphan: true
