+++ .github/workflows/publish-type.yml
@@ -1,0 +1,34 @@
+name: publish Type Package
+on:
+    push:
+        tags:
+            - type-v*
+jobs:
+    Build-and-Publish:
+        runs-on: ubuntu-latest
+        permissions:
+            contents: write
+            packages: write
+        steps:
+            - uses: actions/checkout@v4
+
+            - uses: pnpm/action-setup@v4
+              with:
+                  version: 10
+            - uses: actions/setup-node@v4
+              with:
+                  node-version: 22
+                  registry-url: https://npm.pkg.github.com
+                  cache: pnpm
+            - name: Inject Environment variables
+              run: |
+                  cat > .env <<EOF
+                  ${{ secrets.ENV_FILE }}
+                  EOF
+            - name: Install, Build & Publish
+              run: |
+                  pnpm e i  &&  npm test
+                  cd type/
+                  npm publish
+              env:
+                  NODE_AUTH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
