+++ .vscode/extensions.json
@@ -1,0 +1,18 @@
+{
+    "recommendations": [
+        "yzhang.markdown-all-in-one",
+        "redhat.vscode-yaml",
+        "akamud.vscode-caniuse",
+        "visualstudioexptteam.intellicode-api-usage-examples",
+        "pflannery.vscode-versionlens",
+        "christian-kohler.npm-intellisense",
+        "esbenp.prettier-vscode",
+        "rangav.vscode-thunder-client",
+        "cweijan.vscode-database-client2",
+        "ms-azuretools.vscode-docker",
+        "eamodio.gitlens",
+        "github.vscode-pull-request-github",
+        "github.vscode-github-actions",
+        "github.copilot"
+    ]
+}
