+++ Dockerfile
@@ -1,0 +1,23 @@
+# Reference: https://pnpm.io/docker#example-1-build-a-bundle-in-a-docker-container
+
+FROM node:22-slim AS base
+RUN apt-get update && \
+    apt-get install curl -y --no-install-recommends
+ENV PNPM_HOME="/pnpm"
+ENV PATH="$PNPM_HOME:$PATH"
+RUN npm i pnpm@latest -g
+COPY . /app
+WORKDIR /app
+
+FROM base AS prod-deps
+RUN --mount=type=cache,id=pnpm,target=/pnpm/store  pnpm e i -P --frozen-lockfile
+
+FROM base AS build
+RUN --mount=type=cache,id=pnpm,target=/pnpm/store  pnpm e i --frozen-lockfile
+RUN pnpm build
+
+FROM base
+COPY --from=prod-deps /app/node_modules ./node_modules
+COPY --from=build /app/dist ./dist
+EXPOSE 8080
+CMD ["npm", "start"]
