+++ source/model/Epidemic/index.ts
@@ -1,4 +1,4 @@
-export * from '../Rumor';
+export * from './Rumor';
 export * from './News';
 export * from './AreaDaily';
 export * from './Overall';
